Split TrainingApplications into multiple services to make the classes shorter
