+++ site/app/Admin/Presenters/HomepagePresenter.php
@@ -12,6 +12,7 @@ use MichalSpacekCz\Training\DateList\TrainingApplicationsListFactory;
 use MichalSpacekCz\Training\Dates\TrainingDates;
 use MichalSpacekCz\Training\Dates\UpcomingTrainingDates;
 use MichalSpacekCz\Training\Mails;
+use MichalSpacekCz\Training\Preliminary\PreliminaryTrainings;
 
 class HomepagePresenter extends BasePresenter
 {
@@ -21,6 +22,7 @@ class HomepagePresenter extends BasePresenter
 
 	public function __construct(
 		private readonly TrainingApplications $trainingApplications,
+		private readonly PreliminaryTrainings $trainingPreliminaryApplications,
 		private readonly Mails $trainingMails,
 		private readonly TrainingDates $trainingDates,
 		private readonly UpcomingTrainingDates $upcomingTrainingDates,
@@ -38,7 +40,7 @@ class HomepagePresenter extends BasePresenter
 		$this->template->unpaidInvoices = $this->trainingApplications->getValidUnpaidCount();
 		$this->template->certificates = $certificates = $this->certificates->getNewest();
 		$this->template->certificatesNeedAttention = $this->certsNeedAttention($certificates);
-		[$this->template->preliminaryTotal, $this->template->preliminaryDateSet] = $this->trainingApplications->getPreliminaryCounts();
+		[$this->template->preliminaryTotal, $this->template->preliminaryDateSet] = $this->trainingPreliminaryApplications->getPreliminaryCounts();
 		$this->template->pastWithPersonalData = count($this->trainingDates->getPastWithPersonalData());
 	}
